fix(统计报表-检修）修复数据统计错误bug

+++ aiurt-module-inspection/aiurt-module-inspection-api/src/main/java/com/aiurt/boot/task/dto/OverhaulStatisticsDTOS.java
@@ -71,7 +71,7 @@ public class OverhaulStatisticsDTOS {
     @TableField(exist = false)
     @Excel(name = "总检修时长", width = 15)
     @ApiModelProperty(value = "总检修时长")
-    private Long maintenanceDuration;
+    private float maintenanceDuration;
 
     /**计划检修总数*/
     @TableField(exist = false)
